+++ vite.config.ts
@@ -1,7 +1,6 @@
 import type { ConfigEnv, UserConfigExport } from 'vite';
 import uni from '@dcloudio/vite-plugin-uni';
 import Unocss from 'unocss/vite';
-import { UnoCSSToUni } from 'unocss-preset-uni/vite';
 import { visualizer } from 'rollup-plugin-visualizer';
 import { viteMockServe } from 'vite-plugin-mock';
 
@@ -10,8 +9,6 @@ export default ({ command }: ConfigEnv): UserConfigExport => {
     plugins: [
       uni(),
       Unocss(),
-      // Make sure it's behind Unocss()
-      UnoCSSToUni(),
       visualizer({ emitFile: true }),
       viteMockServe({
         ignore: /^\_/,