+++ CHANGELOG.md
@@ -1,6 +1,41 @@
-# Changelog
+# 更新历史
 
-All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+### 0.0.2 (2022-11-26)
+
+### ♻ Code Refactoring | 代码重构
+
+- 自定义环境变量 TypeScript 智能提示 f566e59
+
+### ⚡ Performance Improvements | 性能优化
+
+- 集成打包体积分析工具 0dcd54c
+- 微信小程序项目设置和优化配置 895469e
+
+### ✨ Features | 新功能
+
+- 安装使用 pinia b929642
+- 安装 luch-request f90b1ff
+- 集成 uni-ui 组件库 cc38610
+- 集成 unocss dbe5d4c
+- 集成 vite-plugin-mock 624670c
+
+### 📝 Documentation | 文档
+
+- 增加项目说明文件 f337cac
+
+### 🐛 Bug Fixes | Bug 修复
+
+- no-async-promise-executor 4a160f7
+
+### 🔧 Continuous Integration | CI 配置
+
+- 更换 unocss 适配 uniapp 依赖 6bb3edd
+- 设置代码提交自动格式化 3302584
+
+### 📦 Chores | 杂物处理
+
+- 更新 uniapp 编译器 a1f3a62
+- **release:** 0.0.1 7c6eb5a
 
 ### 0.0.1 (2022-08-18)